testing armhf support with multiarch/qemu-user-static

+++ makefile
@@ -1,19 +1,25 @@
 # {{{ -- meta
 
-ARCH      := $(shell uname -m | sed "s_armv7l_armhf_")# armhf/x86_64 auto-detect on build
+HOSTARCH  := x86_64#
+QEMUVERS  := v2.9.1-1#
+ARCH      := x86_64# $(shell uname -m | sed "s_armv7l_armhf_")# armhf/x86_64 auto-detect on build
 OPSYS     := alpine
 SHCOMMAND := /bin/bash
 SVCNAME   := base
 USERNAME  := woahbase
 VERSION   := 3.7.0
 
-IMAGETAG  := $(USERNAME)/$(OPSYS)-$(SVCNAME):$(ARCH)
+DOCKEREPO := $(USERNAME)/$(OPSYS)-$(SVCNAME)
+IMAGETAG  := $(DOCKEREPO):$(ARCH)
 
 # -- }}}
 
 # {{{ -- flags
 
-BUILDFLAGS := --rm --force-rm -t $(IMAGETAG) --build-arg VCS_REF=$(shell git rev-parse --short HEAD) --build-arg BUILD_DATE=$(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
+BUILDFLAGS := --rm --force-rm -f $(CURDIR)/Dockerfile_$(ARCH) -t $(IMAGETAG) \
+	--build-arg VCS_REF=$(shell git rev-parse --short HEAD) \
+	--build-arg BUILD_DATE=$(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
+
 CACHEFLAGS := --no-cache=true --pull
 MOUNTFLAGS := #
 NAMEFLAGS  := --name docker_$(SVCNAME) --hostname $(SVCNAME)
@@ -30,11 +36,12 @@ RUNFLAGS   := -c 64 -m 32m # -e PGID=$(shell id -g) -e PUID=$(shell id -u)
 all : run
 
 build : fetch
+	[[ $(ARCH) = "armhf" ]] && (make regbinfmt fetchqemu) || true;
 	docker build $(BUILDFLAGS) $(CACHEFLAGS) $(PROXYFLAGS) .
 
 clean :
-	rm -rf $(CURDIR)/data
-	docker rmi $(IMAGETAG)
+	[[ -d $(CURDIR)/data ]] && rm -rf $(CURDIR)/data || true;
+	docker images | awk '(NR>1) && ($$2!~/none/) {print $$1":"$$2}' | grep $(DOCKEREPO) | xargs -n1 docker rmi
 
 logs :
 	docker logs -f docker_$(SVCNAME)
@@ -43,7 +50,7 @@ pull :
 	docker pull $(IMAGETAG)
 
 push :
-	docker push $(IMAGETAG)
+	docker push $(DOCKEREPO)
 
 restart :
 	docker ps -a | grep 'docker_$(SVCNAME)' -q && docker restart docker_$(SVCNAME) || echo "Service not running.";
@@ -64,16 +71,26 @@ stop :
 	docker stop -t 2 docker_$(SVCNAME)
 
 test :
-	docker run --rm -it $(NAMEFLAGS) $(RUNFLAGS) $(PORTFLAGS) $(MOUNTFLAGS) $(OTHERFLAGS) $(IMAGETAG) bash --version
+	docker run --rm -it $(NAMEFLAGS) $(RUNFLAGS) $(PORTFLAGS) $(MOUNTFLAGS) $(OTHERFLAGS) $(IMAGETAG) -ec "uname -a;"
 
 # -- }}}
 
 # {{{ -- other targets
 
+regbinfmt :
+	docker run --rm --privileged multiarch/qemu-user-static:register --reset
+
 fetch :
 	mkdir -p data && cd data \
-		&& curl \
+	&& curl \
 		-o ./rootfs.tar.gz -SL https://nl.alpinelinux.org/alpine/latest-stable/releases/$(ARCH)/alpine-minirootfs-$(VERSION)-$(ARCH).tar.gz \
 		&& gunzip -f ./rootfs.tar.gz;
+
+fetchqemu :
+	mkdir -p data && \
+	QEMU_ARCH="$$(echo $(ARCH) | sed 's_armhf_arm_')" \
+	&& curl \
+		-o ./data/$(HOSTARCH)_qemu-$${QEMU_ARCH}-static.tar.gz -SL https://github.com/multiarch/qemu-user-static/releases/download/${QEMUVERS}/$(HOSTARCH)_qemu-$${QEMU_ARCH}-static.tar.gz \
+		&& tar xv -C data/ -f ./data/$(HOSTARCH)_qemu-$${QEMU_ARCH}-static.tar.gz;
 
 # -- }}}
